requirements.txt
# requirements.txt
networkx==2.8.4
numpy==1.21.5
pandas==1.4.4
matplotlib==3.5.2
seaborn==0.11.2
scikit-learn==1.0.2
pyyaml==6.0
psutil==5.9.0
